requirements.txt
Django==2.0.3
django-bootstrap4==0.0.6
django-excel-response==2.0.3
et-xmlfile==1.0.1
jdcal==1.3
openpyxl==2.4.11
pytz==2018.3
six==1.11.0
